requirements.txt
aiofiles==0.4.0
aiohttp==3.4.4
alembic==1.3.2
appnope==0.1.0
asn1crypto==0.24.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
backcall==0.1.0
bcolz==0.12.1
Bottleneck==1.2.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
contextlib2==0.6.0.post1
cryptography==2.3.1
cryptography-vectors==2.3.1
cyordereddict==1.0.0
Cython==0.28.5
decorator==4.4.1
empyrical==0.5.0
idna==2.7
idna-ssl==1.1.0
intervaltree==3.0.2
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.12.1
Logbook==0.12.5
lru-dict==1.1.4
Mako==1.1.0
MarkupSafe==1.0
multidict==4.4.2
multipledispatch==0.6.0
nest-asyncio==1.2.1
networkx==1.11
numexpr==2.6.6
numpy==1.11.3
pandas==0.22.0
pandas-datareader==0.6.0
parso==0.5.2
patsy==0.5.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
ptyprocess==0.6.0
py==1.8.1
pycparser==2.19
Pygments==2.5.2
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
PySocks==1.6.8
pyspark==2.4.3
pytest==3.2.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyYAML==3.13
requests==2.20.1
requests-file==1.4.3
requests-ftp==0.3.1
scipy==0.17.1
simplegeneric==0.8.1
six==1.11.0
sortedcontainers==2.1.0
SQLAlchemy==1.2.12
statsmodels==0.9.0
tables==3.4.2
toolz==0.10.0
trading-calendars==1.11.1
traitlets==4.3.2
urllib3==1.23
watermark==2.0.2
wcwidth==0.1.8
yarl==1.2.6
zipline==1.3.0
